Moved navbar to swiping folder and removed dummy accounts. updated imports
