+++ Swiping/src/Navbar/Navbar.css
