+++ Swiping/src/Navbar/testPage.js
